requirements.txt
setuptools~=65.4.1
ete3~=3.1.2
numpy~=1.23.3
matplotlib~=3.6.1
pytest~=7.1.3
testfixtures==6.18.5
pandas~=1.5.0
seaborn~=0.12.0
arviz~=0.12.0
xarray~=2022.3.0
rpy2==3.5.2
scikit-learn~=1.1.1
networkx==2.8.3
scipy==1.8.1
multiprocess-shared-numpy-arrays==-e git+https://github.com/widmi/multiprocess-shared-numpy-arrays
Pillow~=9.2.0